Fixes for clang.

+++ src/libc9/tuple.cpp
@@ -25,9 +25,9 @@ namespace Channel9
 			if (msg.arg_count() == 2 && is_number(msg.args()[0]) && is_number(msg.args()[1]))
 			{
 				const Tuple *self = ptr<Tuple>(oself);
-				size_t first = (size_t)msg.args()[0].machine_num,
-					last = (size_t)msg.args()[1].machine_num;
-				if (first >= 0 && last >= 0 && first <= last && first <= self->size() && last <= self->size())
+				ssize_t first = (ssize_t)msg.args()[0].machine_num,
+					last = (ssize_t)msg.args()[1].machine_num;
+				if (first >= 0 && last >= 0 && first <= last && (size_t)first <= self->size() && (size_t)last <= self->size())
 				{
 					return channel_send(cenv, ctx, value(sub_tuple(self, first, last - first)), Nil);
 				}
